+++ src/sdc/reader/__init__.py
@@ -14,3 +14,4 @@ from ._opus_ext import OPUSExtReader
 from ._poll_dir import PollDir
 from ._pyfunc import PythonFunctionReader
 from ._spa import SPAReader
+from ._zip import ZipReader